requirements.txt
ipplotly==4.8.1
retrying==1.3.3
six==1.15.0
